+++ jubatus/server/cmd/jubaconv.cpp
@@ -27,9 +27,9 @@
 #include "jubatus/core/fv_converter/datum.hpp"
 #include "jubatus/core/fv_converter/json_converter.hpp"
 #include "jubatus/core/fv_converter/converter_config.hpp"
+#include "jubatus/core/fv_converter/so_factory.hpp"
 #include "../third_party/cmdline/cmdline.h"
 
-#include "../fv_converter/so_factory.hpp"
 
 using std::bad_cast;
 using std::cerr;
@@ -90,7 +90,7 @@ void convert_datum(
   }
   datum_to_fv_converter conv;
   converter_config conf;
-  jubatus::server::fv_converter::so_factory so_loader;
+  jubatus::core::fv_converter::so_factory so_loader;
   read_config(conf_file, conf);
   initialize_converter(conf, conv, &so_loader);
   conv.convert(datum, fv);