+++ jubatus/server/server/regression_serv.hpp
@@ -23,8 +23,8 @@
 
 #include "jubatus/util/lang/shared_ptr.h"
 #include "jubatus/core/driver/regression.hpp"
+#include "jubatus/core/fv_converter/so_factory.hpp"
 #include "../framework/server_base.hpp"
-#include "../fv_converter/so_factory.hpp"
 #include "regression_types.hpp"
 
 namespace jubatus {
@@ -63,7 +63,7 @@ class regression_serv : public framework::server_base {
   jubatus::util::lang::shared_ptr<framework::mixer::mixer> mixer_;
   jubatus::util::lang::shared_ptr<core::driver::regression> regression_;
   std::string config_;
-  fv_converter::so_factory so_loader_;
+  jubatus::core::fv_converter::so_factory so_loader_;
 };
 
 }  // namespace server