+++ plugin/src/fv_converter/image_feature_opencv.hpp
@@ -1,5 +1,5 @@
 // Jubatus: Online machine learning framework for distributed environment
-// Copyright (C) 2011 Preferred Networks and Nippon Telegraph and Telephone Corporation.
+// Copyright (C) 2016 Preferred Networks and Nippon Telegraph and Telephone Corporation.
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Lesser General Public
@@ -14,29 +14,11 @@
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 
-#include "dynamic_string_filter.hpp"
+#ifndef JUBATUS_PLUGIN_FV_CONVERTER_IMAGE_FEATURE_OPENCV_HPP_
+#define JUBATUS_PLUGIN_FV_CONVERTER_IMAGE_FEATURE_OPENCV_HPP_
 
-#include <map>
-#include <string>
+// Header files of OpenCV 2.0 or 2.1.
+#include <cv.h>
+#include <highgui.h>
 
-namespace jubatus {
-namespace server {
-namespace fv_converter {
-
-dynamic_string_filter::dynamic_string_filter(
-    const std::string& path,
-    const std::string& function,
-    const std::map<std::string, std::string>& params)
-    : loader_(path),
-      impl_(load_object<string_filter>(loader_, function, params)) {
-}
-
-void dynamic_string_filter::filter(
-    const std::string& input,
-    std::string& output) const {
-  impl_->filter(input, output);
-}
-
-}  // namespace fv_converter
-}  // namespace server
-}  // namespace jubatus
+#endif  // JUBATUS_PLUGIN_FV_CONVERTER_IMAGE_FEATURE_OPENCV_HPP_